Sync main

+++ scripts/generate-browser-readme.js
@@ -96,7 +96,14 @@ fs.readFile(changeLogPath, (err, data) => {
     console.error(err)
   }
 
-  const updatedChangeLog = `${data.toString()} \n${generateNewChangeVersion}`
+  const imageExists = data.includes(generateNewChangeVersion)
+
+  if (imageExists) {
+    console.log(`Image already exists in README and CHANGELOG.`)
+    process.exit(1)
+  }
+
+  const updatedChangeLog = `${data.toString()}${generateNewChangeVersion}`
   fs.writeFileSync(changeLogPath, updatedChangeLog.trim() + "\n", "utf8")
   console.log("Saved CHANGELOG.md at %s", changeLogPath)
